+++ app/fachada/BancAndes.java
@@ -145,7 +145,7 @@ public class BancAndes
 		daoCuentas.registrarCuenta(id, tipo, idCliente);
 	}
 
-	public void cerrarCuenta(int id) throws Exception
+	public void cerrarCuenta(long id) throws Exception
 	{
 		daoCuentas.cerrarCuenta(id);
 	}
